Fix -Wconversion error

+++ ft_lstsize_bonus.c
@@ -6,7 +6,7 @@
 /*   By: sbos <sbos@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/10/11 12:40:50 by sbos          #+#    #+#                 */
-/*   Updated: 2021/11/22 14:25:06 by sbos          ########   odam.nl         */
+/*   Updated: 2021/11/22 18:38:23 by sbos          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
  */
 int	ft_lstsize(t_list *lst)
 {
-	size_t	size;
+	int	size;
 
 	size = 0;
 	while (lst != NULL)
